Split out character descriptions
